pyproject.toml
[tool.poetry]
name = "fastapiproject7"
version = "0.1.0"
description = ""
authors = ["heet.mehta_everestek <heet.mehta@everestek.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.10.4"
boto3 = "^1.21.41"
toml = "^0.10.2"
pathlib = "^1.0.1"
gunicorn = "^20.1.0"
SQLAlchemy = "<2.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.0"
psycopg2 = "^2.9.3"
tomlkit = "^0.10.1"
requests = "^2.28.1"
psycopg2-binary = "^2.9.5"
opentelemetry-api = "^1.15.0"
opentelemetry-instrumentation = "^0.36b0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-instrumentation-fastapi = "^0.36b0"
opentelemetry-instrumentation-requests = "^0.36b0"
opentelemetry-ext-honeycomb = "0.16b0"
loguru = "^0.7.0"
httpx = "^0.24.0"
pytest = "^7.3.1"
coverage = "^7.2.5"
pytest-mock = "^3.10.0"
pytest-html = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
